# swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CryptoPill",
		"contact": {},
		"version": "1.0",
		"description": "API Docs"
	},
	"servers": [
		{
			"url": "http://180.149.241.208:3048",
			"variables": {}
		},
		{
			"url": "http://localhost:3044",
			"variables": {}
		}
	],
	"paths": {
...
		"/user/forgotPassword/{ph_number}": {
			"get": {
				"tags": ["User"],
				"summary": "ForgotPassword",
				"operationId": "ForgotPassword",
				"parameters": [
					{
						"name": "ph_number",
						"in": "path",
						"description": "",
						"required": true,
						"style": "simple",
						"schema": {
							"type": "string",
							"example": "+919594521511"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/user/login": {
			"post": {
				"tags": ["User"],
				"summary": "Step 4: Login",
				"operationId": "Login",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							},
							"example": {
								"contact_number": "+918178426170",
								"password": "12345678"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponse"
								},
								"example": {
									"success": true,
									"data": {
										"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
										"default_role": 2,
										"id": 1,
										"isSetupComplete": 1
									},
									"message": "Success"
								}
							}
						}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/user/changePassword": {
			"post": {
				"tags": ["User"],
				"summary": "Change Password",
				"operationId": "ChangePassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest"
							},
							"example": {
								"current_password": "krishank1234",
								"new_password": "krishank@123",
								"confirm_new_password": "krishank@123"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false
			}
		},
...
		"/doctor/dashboard": {
			"get": {
				"tags": ["Doctor"],
				"summary": "Dashboard",
				"operationId": "Dashboard",
				"parameters": [
					{
						"name": "date",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "string",
							"example": "2020-10-15"
						}
					},
					{
						"name": "day",
						"in": "query",
						"description": "",
						"required": true,
						"style": "form",
						"explode": true,
						"schema": {
							"type": "string",
							"example": "Monday"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Dashboard"
								},
								"example": {
									"success": true,
									"data": [
										{
											"workplace_id": 1,
											"workplace_name": "Sunny11 Hospptal",
											"time_per_appointment": 20,
											"slots": [
												{
													"id": 1,
													"start_time": "10:00:00",
													"end_time": "12:00:00",
													"blocked_slots": []
												}
											],
											"patient_details": []
										},
										{
											"workplace_id": 2,
											"workplace_name": "AIIMS",
											"time_per_appointment": 30,
											"slots": [
												{
													"id": 3,
													"start_time": "15:00:00",
													"end_time": "17:00:00",
													"blocked_slots": []
												}
											],
											"patient_details": []
										}
									],
									"message": "Success"
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
		"/admin/getAllPatientAppointments": {
			"get": {
			  "tags": [
				"Doctor & Staff Management"
			  ],
			  "summary": "getAllPatientAppointments",
			  "operationId": "getAllPatientAppointments",
			  "parameters": [
				{
				  "name": "status",
				  "in": "query",
				  "description": "",
				  "required": true,
				  "style": "form",
				  "explode": true,
				  "schema": {
					"type": "string",
					"example": "Accepted"
				  }
				},
				{
				  "name": "limit",
				  "in": "query",
				  "description": "",
				  "required": true,
				  "style": "form",
				  "explode": true,
				  "schema": {
					"type": "integer",
					"format": "int32",
					"example": 5
				  }
				},
				{
				  "name": "offset",
				  "in": "query",
				  "description": "",
				  "required": true,
				  "style": "form",
				  "explode": true,
				  "schema": {
					"type": "integer",
					"format": "int32",
					"example": 0
				  }
				}
			  ],
			  "responses": {
				"200": {
				  "description": "",
				  "headers": {}
				}
			  },
			  "deprecated": false,
			  "security": []
			}
		},
		"/admin/login": {
			"post": {
				"tags": ["Admin"],
				"summary": "Admin login",
				"operationId": "Adminlogin",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminloginRequest"
							},
							"example": {
								"email": "sunnynpardeshi@gmail.com",
								"password": "Cryptopill12345",
								"role_id": 1
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
		"/admin/forgotPassword": {
			"post": {
				"tags": ["Admin"],
				"summary": "Admin Forgot password",
				"operationId": "AdminForgotpassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminForgotpasswordRequest"
							},
							"example": {
								"email": "sunnynpardeshi@gmail.com",
								"contact_number": "+919594521511"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/admin/changePassword": {
			"post": {
				"tags": ["Admin"],
				"summary": "Admin change password",
				"operationId": "Adminchangepassword",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminchangepasswordRequest"
							},
							"example": {
								"email": "email",
								"password": "password"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
		"/admin/signup": {
			"post": {
				"tags": ["Admin"],
				"summary": "Admin Signup",
				"operationId": "AdminSignup",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminSignupRequest"
							},
							"example": {
								"email": "krishank.tripathi@neosoftmail.com",
								"contact_number": "+918128153228",
								"first_name": "Krishank",
								"last_name": "Tripathi"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {}
					}
				},
				"deprecated": false,
				"security": []
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"LoginRequest": {
				"title": "Login Request",
				"required": ["contact_number", "password"],
				"type": "object",
				"properties": {
					"contact_number": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"example": {
					"contact_number": "+918178426170",
					"password": "12345678"
				}
			},
			"LoginResponse": {
				"title": "Login Response",
				"required": ["success", "data", "message"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/Data2"
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"success": true,
					"data": {
						"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
						"default_role": 2,
						"id": 1,
						"isSetupComplete": 1
					},
					"message": "Success"
				}
			},
			"Data2": {
				"title": "Data2",
				"required": ["token", "default_role", "id", "isSetupComplete"],
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					},
					"default_role": {
						"type": "integer",
						"format": "int32"
					},
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"isSetupComplete": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZmlyc3RfbmFtZSI6InNhdGlzaCIsIm1pZGRsZV9uYW1lIjoiIiwibGFzdF9uYW1lIjoiam9zaGkiLCJlbWFpbCI6ImR1bW15ZG9jQGdtYWlsLmNvbSIsImdlbmRlciI6Im1hbGUiLCJiaXJ0aF9kYXRlIjoiMTk4NS0wOC0wMSIsInByb2ZpbGVfaW1hZ2UiOiJodHRwczovL2NyeXB0b3BpbGxhcHAuczMuYW1hem9uYXdzLmNvbS9kb2N0b3IvMS9kZW1vLmpwZyIsIm5ld19wcm9maWxlX2ltYWdlIjpudWxsLCJjb250YWN0X251bWJlciI6Iis5MTgxNzg0MjYxNzAiLCJhY2NvdW50X2FjdGl2YXRpb24iOjEsInByb2ZpbGVfaW1hZ2VfdmVyaWZ5IjowLCJsYWJfb3JfcGhhcm1hX2VtcGxveWVtZW50X251bWJlciI6bnVsbCwiZW1haWxfdmVyaWZ5IjowLCJwaG9uZV92ZXJpZnkiOjEsImRlZmF1bHRfcm9sZSI6MiwiaXNBZG1pbiI6MCwiaXNTZXR1cENvbXBsZXRlIjoxLCJkZXZpY2VUb2tlbiI6bnVsbCwiZGV2aWNlVHlwZSI6bnVsbCwicm9sZV9pZCI6Miwicm9sZSI6ImRvY3RvciIsImlzV29ya3BsYWNlQWRtaW4iOjEsImlhdCI6MTYwNTY4MDA5MSwiZXhwIjoxNjA2OTc2MDkxfQ.x35_SNQCcEIRHt2_fmj-8-hxZ3YGeyiLRTt9B87sGrk",
					"default_role": 2,
					"id": 1,
					"isSetupComplete": 1
				}
			},
...
			"ChangePasswordRequest": {
				"title": "ChangePasswordRequest",
				"required": [
					"current_password",
					"new_password",
					"confirm_new_password"
				],
				"type": "object",
				"properties": {
					"current_password": {
						"type": "string"
					},
					"new_password": {
						"type": "string"
					},
					"confirm_new_password": {
						"type": "string"
					}
				},
				"example": {
					"current_password": "krishank1234",
					"new_password": "krishank@123",
					"confirm_new_password": "krishank@123"
				}
			},
...
			"Dashboard": {
				"title": "Dashboard",
				"required": ["success", "data", "message"],
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Data5"
						},
						"description": ""
					},
					"message": {
						"type": "string"
					}
				},
				"example": {
					"success": true,
					"data": [
						{
							"workplace_id": 1,
							"workplace_name": "Sunny11 Hospptal",
							"time_per_appointment": 20,
							"slots": [
								{
									"id": 1,
									"start_time": "10:00:00",
									"end_time": "12:00:00",
									"blocked_slots": []
								}
							],
							"patient_details": []
						},
						{
							"workplace_id": 2,
							"workplace_name": "AIIMS",
							"time_per_appointment": 30,
							"slots": [
								{
									"id": 3,
									"start_time": "15:00:00",
									"end_time": "17:00:00",
									"blocked_slots": []
								}
							],
							"patient_details": []
						}
					],
					"message": "Success"
				}
			},
			"Data5": {
				"title": "Data5",
				"required": [
					"workplace_id",
					"workplace_name",
					"time_per_appointment",
					"slots",
					"patient_details"
				],
				"type": "object",
				"properties": {
					"workplace_id": {
						"type": "integer",
						"format": "int32"
					},
					"workplace_name": {
						"type": "string"
					},
					"time_per_appointment": {
						"type": "integer",
						"format": "int32"
					},
					"slots": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Slot1"
						},
						"description": ""
					},
					"patient_details": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"workplace_id": 1,
					"workplace_name": "Sunny11 Hospptal",
					"time_per_appointment": 20,
					"slots": [
						{
							"id": 1,
							"start_time": "10:00:00",
							"end_time": "12:00:00",
							"blocked_slots": []
						}
					],
					"patient_details": []
				}
			},
			"Slot1": {
				"title": "Slot1",
				"required": ["id", "start_time", "end_time", "blocked_slots"],
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"start_time": {
						"type": "string"
					},
					"end_time": {
						"type": "string"
					},
					"blocked_slots": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": ""
					}
				},
				"example": {
					"id": 1,
					"start_time": "10:00:00",
					"end_time": "12:00:00",
					"blocked_slots": []
				}
			},
...
			"AdminloginRequest": {
				"title": "AdminloginRequest",
				"required": ["email", "password", "role_id"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role_id": {
						"type": "integer",
						"format": "int32"
					}
				},
				"example": {
					"email": "sunnynpardeshi@gmail.com",
					"password": "Cryptopill12345",
					"role_id": 1
				}
			},
			"AdminForgotpasswordRequest": {
				"title": "AdminForgotpasswordRequest",
				"required": ["email", "contact_number"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					}
				},
				"example": {
					"email": "sunnynpardeshi@gmail.com",
					"contact_number": "+919594521511"
				}
			},
			"AdminchangepasswordRequest": {
				"title": "AdminchangepasswordRequest",
				"required": ["email", "password"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"example": {
					"email": "email",
					"password": "password"
				}
			},
			"AdminSignupRequest": {
				"title": "AdminSignupRequest",
				"required": ["email", "contact_number", "first_name", "last_name"],
				"type": "object",
				"properties": {
					"email": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"example": {
					"email": "krishank.tripathi@neosoftmail.com",
					"contact_number": "+918128153228",
					"first_name": "Krishank",
					"last_name": "Tripathi"
				}
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"security": [
		{
			"ApiKeyAuth": []
		}
	]
}